Minor changes
I'm wondering if you want to include some verbiage in the first section, after the second paragraph, with more details about whether the user *should* be configuring their VMs to create a memory dump or just crash. Offering a suggestion (sounds like the suggestion would be to configure a VM to create a memory dump?) to the user might help out folks unfamiliar NMI/SysRq.